Checking refactoring of the parser API.
Now we have a single input circular buffer on the decoders,
and an generic parser that behaves like the original TI 
circular buffer.
The new circular buffer includes logic to preserve the buffers
metadata, allowing the optimized parsers to work as well.
The specialized parsers may indicate their memory requirements
to be more efficient than the generic parser.
Implementing as well the audio decoding capabilities and providing
an aac parser



git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/branches/BRANCH_DDOMPE/gstreamer_ti/ti_build/ticodecplugin@434 24075187-2e39-4e88-bbb8-bc8aa768f540